+++ docs/src/Types.fs
@@ -21,6 +21,7 @@ type Msg =
     | CardMsg of Components.Card.Types.Msg
     | MediaMsg of Components.Media.Types.Msg
     | MenuMsg of Components.Menu.Types.Msg
+    | MessageMsg of Components.Message.Types.Msg
 
 type ElementsModel =
     { Box : Elements.Box.Types.Model
@@ -40,7 +41,8 @@ type ComponentsModel =
       Breadcrumb : Components.Breadcrumb.Types.Model
       Card : Components.Card.Types.Model
       Media : Components.Media.Types.Model
-      Menu : Components.Menu.Types.Model }
+      Menu : Components.Menu.Types.Model
+      Message : Components.Message.Types.Model }
 
 type Model =
     { CurrentPage : Page