+++ RELEASE_NOTES.md
@@ -1,3 +1,9 @@
+### 0.1.0-alpha10
+
+* Fix `Message` colors properties
+* Add `Message` documentation
+* Add `Size` properties to `Message`
+
 ### 0.1.0-alpha9
 
 * Add `Card` documentation